move compute_vwind_impl_wgt into metric_fields.py
